candidate v2.0.0 integrating both icarus dev bugfix and getting ready for unified variables by integrating prohect SBN dev

+++ profit/pylib.py
@@ -46,7 +46,7 @@ class DataFrameFormula(object):
 
         ret.name = self.name
 
-        return ret
+        return ret.astype(np.float32)
 
     def PrintValue(self):
         return self.formula
